curriculum working well
