+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "firebase": "^12.1.0",
+    "firebase-functions": "^6.4.0",
     "js-cookie": "^3.0.5",
     "moment-timezone": "^0.6.0",
     "next": "15.4.6",
