+++ src/components/DataTable.tsx
@@ -32,6 +32,7 @@ import {
 import { JobPosting } from '@/types/job.types';
 import JobDetailModal from './JobDetailModal';
 import { useAuth } from './auth/AuthContext';
+import { fetchJobText } from '@/lib/features/textJob/textJobTruck';
 
 export default function DataTable() {
   const dispatch = useAppDispatch();
@@ -144,6 +145,7 @@ export default function DataTable() {
         appliedbyme={job.appliedbyme}
         ignore={job.ignore}
         wait={job.wait}
+        close={job.closed}
       >
         <JobCardHeader
           title={
