+++ src/components/JobDetailModal.tsx
@@ -1,7 +1,9 @@
 'use client';
 
-import React, { useState } from 'react';
-import { Button, Typography, Tooltip, IconButton, Box } from '@mui/material';
+import React, { use, useState } from 'react';
+
+import { getFunctions, httpsCallable } from "firebase/functions";
+import { Button, Typography, Tooltip, IconButton, Box, CircularProgress } from '@mui/material';
 import {
   Close as CloseIcon,
   OpenInNew as OpenInNewIcon,
@@ -9,7 +11,8 @@ import {
   HourglassEmpty as HourglassEmptyIcon,
   ThumbUp as ThumbUpIcon,
   Lock as LockIcon,
-  ContentCopy as ContentCopyIcon
+  ContentCopy as ContentCopyIcon,
+  ArrowForward as ArrowForwardIcon
 } from '@mui/icons-material';
 import { JobPosting } from '@/types/job.types';
 import JobDescription from './JobDescription';
@@ -29,6 +32,8 @@ import {
   StyledDialogActions,
   ActionButtonsGroup
 } from './JobDetailModal.styled';
+import { useAuth } from './auth/AuthContext';
+import { appFirebase, functions } from './auth/firebaseConfig';
 
 interface JobDetailModalProps {
   job: JobPosting | null;
@@ -49,9 +54,11 @@ export default function JobDetailModal({ job, open, onClose,
   handleLockAction }: JobDetailModalProps) {
 
   const [copySuccess, setCopySuccess] = useState(false);
+  const { user, getAuthToken } = useAuth();
+  const [loading, setLoading] = useState(false);
 
   // Get job description from Redux store
-  const jobDescription = useAppSelector(state => state.textJob.descriptions[job?._id || '']);
+  const jobDescription = useAppSelector(state => state.textJob.descriptions);
 
   if (!job) return null;
 
@@ -64,6 +71,29 @@ export default function JobDetailModal({ job, open, onClose,
       setTimeout(() => setCopySuccess(false), 2000); // Hide success message after 2 seconds
     } catch (err) {
       console.error('Failed to copy text: ', err);
+    }
+  };
+
+
+  const handleOpenCV = async () => {
+    setLoading(true);
+    const generateCv = httpsCallable(functions, "generateCv");
+    try {
+      const token = await getAuthToken();
+      const result = await generateCv({
+        jobDescription: job.description,
+        authToken: token
+      });
+      const cvData = result.data;
+      // Store CV data and opportunity ID in local storage to pass to the CV page
+      localStorage.setItem("cvData", JSON.stringify(cvData));
+      localStorage.setItem("opportunityId", job._id);
+      window.open('/cv', '_blank');
+    } catch (error) {
+      console.error("Error generating CV:", error);
+      alert("Error generating CV. Please try again.");
+    } finally {
+      setLoading(false);
     }
   };
 
@@ -207,20 +237,57 @@ export default function JobDetailModal({ job, open, onClose,
     </>
   )
 
-  return (
-    <StyledDialog open={open} onClose={onClose} maxWidth="md" fullWidth>
-      <StyledDialogTitle>
-        <Typography variant="h6">{job.title} | {job.workplaceTypes} | {job.country} | {job.formattedEmploymentStatus} | {`AP:${job.applies}`}</Typography>
-        <CloseButton onClick={onClose}>
-          <CloseIcon />
-        </CloseButton>
-      </StyledDialogTitle>
+  const JobDescriptionTag = (
+    <ContentSection>
+      <Box display="flex" alignItems="center" gap={1} mb={2}>
+        <SectionTitle style={{ margin: 0 }}>Job Description</SectionTitle>
+        <Tooltip title="Copy job description">
+          <IconButton
+            onClick={handleCopyDescription}
+            size="small"
+            sx={{
+              color: '#6b7280',
+              '&:hover': { color: '#374151' }
+            }}
+          >
+            <ContentCopyIcon fontSize="small" />
+          </IconButton>
+        </Tooltip>
+        {job._id}
+        {copySuccess && (
+          <Typography
+            variant="caption"
+            sx={{
+              color: '#10b981',
+              fontWeight: 500,
+              ml: 1
+            }}
+          >
+            Copied!
+          </Typography>
+        )}
+      </Box>
+      <JobDescription jobId={job._id} />
+    </ContentSection>
+  )
 
+  const boxLoad = (
+    <Box display="flex" justifyContent="center" alignItems="center" height="100vh">
+        <CircularProgress />
+        BUILDING CURRICULUM VITAE
+      </Box>
+  )
+
+  const contentBody = (
+    <>
+    
       <StyledDialogContent>
         <CompanyHeader>
           <div className="company-name">{job.companyName}</div>
           <div className="job-id">ID: {job._id}</div>
         </CompanyHeader>
+
+
 
         {job.llm && (
           <ContentSection>
@@ -237,39 +304,8 @@ export default function JobDetailModal({ job, open, onClose,
         {JobDetailsInPart}
         {JobSallaryInPart}
 
-        <ContentSection>
-          <Box display="flex" alignItems="center" gap={1} mb={2}>
-            <SectionTitle style={{ margin: 0 }}>Job Description</SectionTitle>
-            <Tooltip title="Copy job description">
-              <IconButton 
-                onClick={handleCopyDescription}
-                size="small"
-                sx={{ 
-                  color: '#6b7280',
-                  '&:hover': { color: '#374151' }
-                }}
-              >
-                <ContentCopyIcon fontSize="small" />
-              </IconButton>
-            </Tooltip>
-            {job._id}
-            {copySuccess && (
-              <Typography 
-                variant="caption" 
-                sx={{ 
-                  color: '#10b981', 
-                  fontWeight: 500,
-                  ml: 1
-                }}
-              >
-                Copied!
-              </Typography>
-            )}
-          </Box>
-          <JobDescription jobId={job._id} />
-        </ContentSection>
 
-
+        {JobDescriptionTag}
 
         <ContentSection>
           <Box display="flex" justifyContent="space-between" alignItems="flex-start" mb={2}>
@@ -282,6 +318,7 @@ export default function JobDetailModal({ job, open, onClose,
 
       </StyledDialogContent>
 
+
       <StyledDialogActions>
         <Button
           variant="outlined"
@@ -292,6 +329,8 @@ export default function JobDetailModal({ job, open, onClose,
         </Button>
 
         <ActionButtonsGroup>
+
+
           <Tooltip title="Open Job">
             <IconButton
               className="go-action"
@@ -300,44 +339,71 @@ export default function JobDetailModal({ job, open, onClose,
               <OpenInNewIcon />
             </IconButton>
           </Tooltip>
+          {user &&
+            <>
+              <Tooltip title="Open CV">
+                <IconButton
+                  onClick={handleOpenCV}
+                  sx={{
+                    color: '#10b981',
+                    '&:hover': {
+                      color: '#059669',
+                      backgroundColor: 'rgba(16, 185, 129, 0.1)'
+                    }
+                  }}
+                >
+                  <ArrowForwardIcon />
+                </IconButton>
+              </Tooltip>
+              <Tooltip title={job.ignore ? "Undo Ignore" : "Ignore Job"}>
+                <IconButton
+                  className="reject-action"
+                  onClick={() => handleRejectAction(job)}
+                >
+                  <ThumbDownIcon />
+                </IconButton>
+              </Tooltip>
 
-          <Tooltip title={job.ignore ? "Undo Ignore" : "Ignore Job"}>
-            <IconButton
-              className="reject-action"
-              onClick={() => handleRejectAction(job)}
-            >
-              <ThumbDownIcon />
-            </IconButton>
-          </Tooltip>
+              <Tooltip title={job.wait ? "Undo Wait" : "Mark as Waiting"}>
+                <IconButton
+                  className="wait-action"
+                  onClick={() => handleWaitAction(job)}
+                >
+                  <HourglassEmptyIcon />
+                </IconButton>
+              </Tooltip>
 
-          <Tooltip title={job.wait ? "Undo Wait" : "Mark as Waiting"}>
-            <IconButton
-              className="wait-action"
-              onClick={() => handleWaitAction(job)}
-            >
-              <HourglassEmptyIcon />
-            </IconButton>
-          </Tooltip>
+              <Tooltip title={job.appliedbyme ? "Undo Applied" : "Mark as Applied"}>
+                <IconButton
+                  className="accept-action"
+                  onClick={() => handleAcceptAction(job)}
+                >
+                  <ThumbUpIcon />
+                </IconButton>
+              </Tooltip>
 
-          <Tooltip title={job.appliedbyme ? "Undo Applied" : "Mark as Applied"}>
-            <IconButton
-              className="accept-action"
-              onClick={() => handleAcceptAction(job)}
-            >
-              <ThumbUpIcon />
-            </IconButton>
-          </Tooltip>
-
-          <Tooltip title={job.closed ? "Reopen Job" : "Close Job"}>
-            <IconButton
-              className="lock-action"
-              onClick={() => handleLockAction(job)}
-            >
-              <LockIcon />
-            </IconButton>
-          </Tooltip>
+              <Tooltip title={job.closed ? "Reopen Job" : "Close Job"}>
+                <IconButton
+                  className="lock-action"
+                  onClick={() => handleLockAction(job)}
+                >
+                  <LockIcon />
+                </IconButton>
+              </Tooltip>
+            </>}
         </ActionButtonsGroup>
       </StyledDialogActions>
+</>
+  )
+  return (
+    <StyledDialog open={open} onClose={onClose} maxWidth="md" fullWidth>
+      <StyledDialogTitle>
+        <Typography >{job.title} | {job.workplaceTypes} | {job.country} | {job.formattedEmploymentStatus} | {`AP:${job.applies}`}</Typography>
+        <CloseButton onClick={onClose}>
+          <CloseIcon />
+        </CloseButton>
+      </StyledDialogTitle>
+      {loading ? boxLoad : contentBody}
     </StyledDialog>
   );
 }
